C++ compiler log identification was improved. RowLengths was renamed to CompressedRowsLengths. OpenMP support wads added to matrix-vector multiplication and to scalar products. CSR format can use Cusparse for SpMV.

+++ src/core/tnlCudaDeviceInfo.h
@@ -18,6 +18,7 @@
 #ifndef TNLCUDADEVICEINFO_H
 #define	TNLCUDADEVICEINFO_H
 
+#include <stdlib.h>
 #include <core/tnlCuda.h>
 
 class tnlCudaDeviceInfo
@@ -34,7 +35,7 @@ class tnlCudaDeviceInfo
       
       static int getClockRate( int deviceNum );
       
-      static int getGlobalMemory( int deviceNum );
+      static size_t getGlobalMemory( int deviceNum );
 
       static int getMemoryClockRate( int deviceNum );
